Rename _property_testers.py to _asserts.py
