+++ tests/unit/aggregation/test_config.py
@@ -4,12 +4,12 @@ from torch import Tensor
 
 from torchjd.aggregation import ConFIG
 
-from ._inputs import non_strong_matrices, scaled_matrices, typical_matrices
-from ._property_testers import (
+from ._asserts import (
     assert_expected_structure,
     assert_linear_under_scaling,
     assert_non_differentiable,
 )
+from ._inputs import non_strong_matrices, scaled_matrices, typical_matrices
 
 scaled_pairs = [(ConFIG(), matrix) for matrix in scaled_matrices]
 typical_pairs = [(ConFIG(), matrix) for matrix in typical_matrices]
