+++ tests/unit/aggregation/test_upgrad.py
@@ -4,8 +4,7 @@ from torch import Tensor
 
 from torchjd.aggregation import UPGrad
 
-from ._inputs import non_strong_matrices, scaled_matrices, typical_matrices
-from ._property_testers import (
+from ._asserts import (
     assert_expected_structure,
     assert_linear_under_scaling,
     assert_non_conflicting,
@@ -13,6 +12,7 @@ from ._property_testers import (
     assert_permutation_invariant,
     assert_strongly_stationary,
 )
+from ._inputs import non_strong_matrices, scaled_matrices, typical_matrices
 
 scaled_pairs = [(UPGrad(), matrix) for matrix in scaled_matrices]
 typical_pairs = [(UPGrad(), matrix) for matrix in typical_matrices]
